fixed clang++ and g++ c++11 "narrowing convertion" compilation errors

+++ src/PluginAuto/Mac/PluginWindowMacCG.cpp
@@ -110,7 +110,8 @@ bool PluginWindowMacCG::SendEvent(PluginEvent* evt) {
         } else if (rval && evt->validType<MacEventCocoa>()) {
             FB::MacEventCocoa *event = evt->get<MacEventCocoa>();
             if (NPCocoaEventDrawRect == event->msg.type) {
-                FB::Rect bounds = { 0, 0, event->msg.data.draw.height, event->msg.data.draw.width };
+                FB::Rect bounds = { 0, 0, static_cast<int32_t>(event->msg.data.draw.height),
+                                    static_cast<int32_t>(event->msg.data.draw.width) };
                 DrawLabel(event->msg.data.draw.context, bounds);
             }
         } else if (rval && evt->validType<MacEventCarbon>() && rval) {
